No longer make decisions on the partial download marker
We no longer look at the file to see if we have a partial dataset or
not. Removed the method
and also the tests.
Fixes #21

It is a good idea to keep the marker for now - it is a good UI aid.
However, there are no tests, so we will see how long before it breaks.
Fixes #23